requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# NLP and ML libraries
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.5.1
faiss-cpu==1.13.1
numpy>=1.25.0
scikit-learn==1.3.2

# LangChain ecosystem
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
tiktoken==0.5.2

# Advanced NLP Models
openai-whisper==20231117
sentencepiece==0.1.99

# Language detection and translation
langdetect==1.0.9
translators==5.9.1
gtts==2.4.0
deep-translator==1.11.4
google-generativeai==0.3.2

# Audio processing for Whisper
soundfile==0.12.1

# Quantized LLM support (Gemma2-9B Q4)
llama-cpp-python==0.2.90
librosa==0.10.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0

# Additional multilingual packages
# Already included above: openai-whisper, gtts, transformers
# mBERT and XLM-R are available through transformers package